requirements.txt
accelerate==1.11.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
audioread==3.1.0
av==16.0.1
babel==2.17.0
beautifulsoup4==4.14.2
blis==1.3.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
csvw==3.7.0
ctranslate2==4.6.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
diffusers==0.35.1
dlinfo==2.0.0
docopt==0.6.2
ebooklib==0.20
espeakng-loader==0.2.4
executing==2.2.1
faiss-cpu==1.13.0
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib-metadata==8.7.0
ipykernel==6.30.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.9
kokoro==0.9.4
kokoro-onnx==0.3.9
kokoro-tts==2.3.0
langchain==1.1.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-ollama==1.0.0
langchain-text-splitters==1.0.0
langcodes==3.5.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.48
language-data==1.3.0
language-tags==1.2.0
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.45.1
loguru==0.7.3
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
misaki==0.9.4
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.13
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
num2words==0.5.14
numba==0.62.1
numpy==2.3.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.6.1
onnxruntime==1.23.2
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
phonemizer-fork==3.3.1
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
preshed==3.0.10
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.12.0
pygments==2.19.2
pymupdf==1.26.5
pymupdf4llm==0.0.27
pyparsing==3.2.5
pypdf==6.4.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
pyyaml==6.0.3
pyzmq==27.1.0
rdflib==7.4.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.2.0
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
segments==2.3.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.4.1
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.44
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
tenacity==9.1.2
termcolor==3.2.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.2.0
typer==0.20.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
win32-setctime==1.2.0
wrapt==2.0.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
